PID needs to be tuned? Not too sure if the implementation is correct.

+++ Utilities/motor.c
@@ -1,6 +1,6 @@
 #include "motor.h"
 
-void PWM_Config(uint16_t period, uint16_t high)
+void PWM_Config(uint32_t period, uint16_t high)
 {
 	// PF2 Output Initialization
